Add files via upload
Autonomous Collision Avoidance Car

In Engineering 115 Engineering Graphics, I was tasked to build an autonomous car that would drive through a course of a combination of four straights and four turns. I increased the efficiency of this task by adding a method that would allow the car to reverse if it detected a range of zero to ten inches, which allowed it to have higher collision avoidance. This code was written in C/C++ and uploaded to an Arduino unit.