Add files via upload
Stove for Developing Nations

In Engineering 120 Introduction to Engineering, I was tasked with a group to design a stove to minimize energy usage and emission output. I wrote a program in Python to calculate the energy required to cook a meal with inputs from the ingredients' specific heat capacities and densities.